==> code/zebrafish_heart_i001/4A_seurat_integration_lognorm.R <==
library(Seurat)

scrnaseq <- readRDS("./output/processed_data/SCD-VI-i001/zebrafish_heart_scrnaseq_lognorm_pca.rds")
visium_subset <- readRDS("./output/processed_data/SCD-VI-i001/SCD-VI-i001_subset_lognorm_pca_clustered.rds")

# Anchor-based integration
anchors <- FindTransferAnchors(
  reference = scrnaseq, query = visium_subset, k.filter = NA,
  normalization.method = "LogNormalize"
)

visium_subset <- TransferData(
  anchorset = anchors, refdata = scrnaseq$plot.ident2, query = visium_subset,
  prediction.assay = TRUE, weight.reduction = visium_subset[["pca"]], dims = 1:30
)

saveRDS(visium_subset, "./output/processed_data/SCD-VI-i001/SCD-VI-i001_lognorm_seurat_predictions.rds")
